+++ src/fieldsets/ListFieldset.jsx
@@ -92,8 +92,9 @@ class ListFieldset extends React.Component {
       selectField = (
         <FormField
           label={this.props.numLabel}
-          value={this.state.length}
+          value={this.state.length+''}
           onChange={this.handleNumChange}
+          optionLabelProp="children"
         >
           <Select style={{ width: 100 }}>
             {options}