+++ CHANGELOG.md
@@ -1,5 +1,23 @@
 # 更新日志
 
+## 1.1.2
+`2018-01-17`
+- 修复radioFieldset组件。
+- 修改DateTimeInput 输入控件。
+
+## 1.1.1
+`2018-01-17`
+- 升级fieldsets组件。
+- GroupFieldset组件与CollapseFieldset组件合并。
+- 增加fieldset 分块提交。
+
+## 1.1.0
+`2018-01-16`
+- 增加字段转化组件，优化自动生成方案。
+- 使 controls 和 fieldsets 可扩展。
+- 增加Select control 和 多种fieldset组件。
+
+
 ## 1.0.0
 `2017-12-04`
 - 自动生成表单组件包含9种基础输入控件。